spider/requirements.txt
markitdown
scrapy
flask
pymongo
urllib3
requests
python-dotenv
asyncio
aiohttp
qdrant-client